backend/requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==8.3.1
promise==2.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==1.7.1
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
